HV-1363 Some editorialization
