+++ engine/src/main/java/org/hibernate/validator/spi/properties/package-info.java
@@ -6,7 +6,7 @@
  */
 
 /**
- * <p>This package provides support for customization of the JavaBeans property detection logic.</p>
+ * <p>This package provides support for customizing the getter detection logic.</p>
  * <p>This package is part of the public Hibernate Validator SPI.</p>
  */
 package org.hibernate.validator.spi.properties;
